+++ CHANGELOG.md
@@ -10,6 +10,7 @@ We follow [Semantic Versions](https://semver.org/).
   - renamed `helpers.enter` to `helpers.resolve`
   - deleted `dependency`, `init_dependencies`, `shutdown_dependencies` in favor of
     `registry.set_scope`, `registry.init`, `registry.shutdown`
+  - deleted `helpers.lifespan` in favor of `registry.lifespan` and `registry.alifespan`
 
 ## Version 0.31.0